refactor: extract checkbox logic to reusable CheckboxMixin
