+++ packages/checkbox/src/vaadin-checkbox.js
@@ -4,16 +4,10 @@
  * This program is available under Apache License Version 2.0, available at https://vaadin.com/license/
  */
 import { html, PolymerElement } from '@polymer/polymer/polymer-element.js';
-import { ActiveMixin } from '@vaadin/component-base/src/active-mixin.js';
-import { ControllerMixin } from '@vaadin/component-base/src/controller-mixin.js';
-import { DelegateFocusMixin } from '@vaadin/component-base/src/delegate-focus-mixin.js';
 import { ElementMixin } from '@vaadin/component-base/src/element-mixin.js';
 import { TooltipController } from '@vaadin/component-base/src/tooltip-controller.js';
-import { CheckedMixin } from '@vaadin/field-base/src/checked-mixin.js';
-import { InputController } from '@vaadin/field-base/src/input-controller.js';
-import { LabelMixin } from '@vaadin/field-base/src/label-mixin.js';
-import { LabelledInputController } from '@vaadin/field-base/src/labelled-input-controller.js';
 import { registerStyles, ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
+import { CheckboxMixin } from './vaadin-checkbox-mixin.js';
 import { checkboxStyles } from './vaadin-checkbox-styles.js';
 
 registerStyles('vaadin-checkbox', checkboxStyles, { moduleId: 'vaadin-checkbox-styles' });
@@ -51,17 +45,11 @@ registerStyles('vaadin-checkbox', checkboxStyles, { moduleId: 'vaadin-checkbox-s
  * @fires {CustomEvent} indeterminate-changed - Fired when the `indeterminate` property changes.
  *
  * @extends HTMLElement
- * @mixes ControllerMixin
+ * @mixes CheckboxMixin
  * @mixes ThemableMixin
  * @mixes ElementMixin
- * @mixes ActiveMixin
- * @mixes DelegateFocusMixin
- * @mixes CheckedMixin
- * @mixes LabelMixin
  */
-class Checkbox extends LabelMixin(
-  CheckedMixin(DelegateFocusMixin(ActiveMixin(ElementMixin(ThemableMixin(ControllerMixin(PolymerElement)))))),
-) {
+export class Checkbox extends CheckboxMixin(ElementMixin(ThemableMixin(PolymerElement))) {
   static get is() {
     return 'vaadin-checkbox';
   }
@@ -77,107 +65,13 @@ class Checkbox extends LabelMixin(
     `;
   }
 
-  static get properties() {
-    return {
-      /**
-       * True if the checkbox is in the indeterminate state which means
-       * it is not possible to say whether it is checked or unchecked.
-       * The state is reset once the user switches the checkbox by hand.
-       *
-       * https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/checkbox#Indeterminate_state_checkboxes
-       *
-       * @type {boolean}
-       */
-      indeterminate: {
-        type: Boolean,
-        notify: true,
-        value: false,
-        reflectToAttribute: true,
-      },
-
-      /**
-       * The name of the checkbox.
-       *
-       * @type {string}
-       */
-      name: {
-        type: String,
-        value: '',
-      },
-    };
-  }
-
-  /** @override */
-  static get delegateProps() {
-    return [...super.delegateProps, 'indeterminate'];
-  }
-
-  /** @override */
-  static get delegateAttrs() {
-    return [...super.delegateAttrs, 'name'];
-  }
-
-  constructor() {
-    super();
-
-    this._setType('checkbox');
-
-    // Set the string "on" as the default value for the checkbox following the HTML specification:
-    // https://html.spec.whatwg.org/multipage/input.html#dom-input-value-default-on
-    this.value = 'on';
-  }
-
   /** @protected */
   ready() {
     super.ready();
 
-    this.addController(
-      new InputController(this, (input) => {
-        this._setInputElement(input);
-        this._setFocusElement(input);
-        this.stateTarget = input;
-        this.ariaTarget = input;
-      }),
-    );
-    this.addController(new LabelledInputController(this.inputElement, this._labelController));
     this._tooltipController = new TooltipController(this);
     this.addController(this._tooltipController);
-  }
-
-  /**
-   * Extends the method from `ActiveMixin` in order to
-   * prevent setting the `active` attribute when interacting with a link inside the label.
-   *
-   * @param {Event} event
-   * @return {boolean}
-   * @protected
-   * @override
-   */
-  _shouldSetActive(event) {
-    if (event.target.localName === 'a') {
-      return false;
-    }
-
-    return super._shouldSetActive(event);
-  }
-
-  /**
-   * Extends the method from `CheckedMixin` in order to
-   * reset the indeterminate state once the user switches the checkbox.
-   *
-   * @param {boolean} checked
-   * @protected
-   * @override
-   */
-  _toggleChecked(checked) {
-    if (this.indeterminate) {
-      this.indeterminate = false;
-    }
-
-    super._toggleChecked(checked);
   }
 }
 
 customElements.define(Checkbox.is, Checkbox);
-
-export { Checkbox };
